Walk Through Walls and Reroll Map
The Walk Through Walls cheat is now enabled for the World Map (hit F12). This includes a little HUD note to say it's on or not.

Also, you can reroll the map with F5, and it properly clears the old map and rolls a whole new one. (You might end up in water if this happens right now, so feel free to use the Walk Through Walls cheat to fix that!)