Regions now function properly!
Got them working. The holdup was silly and obvious with a fresh head:  The command for "hey, that tile you just added to the region, add it to the list of tiles on the region's edge" was indented one tab too far - which meant it was only happening if the region's count was zero, if it was the very first tile being added.